data/preprocessing/gpt/requirements.txt
openai
python-dotenv
tenacity
